requirements.txt
PySide6
numpy
pillow
matplotlib
fabio
